requirements.txt
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-model-utils==3.1.2
djangorestframework==3.8.2
djangorestframework-stubs==1.2.0
drf-yasg==1.17.1
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
psycopg2-binary==2.8.4
pyparsing==2.4.7
pytz==2019.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.9
